Filtered to active users
Added a filter to the Get-ADUser command to limit the results to active accounts, ignoring any that are disabled or have a password that doesn't expire.

Also changed the output to the Downloads folder, and deleted the get-emails.ps1 script as it is no longer needed (functionality rolled into get-users.ps1).