system_archive v1.2
Update 1.2 adds the option to back up both the system registry, as well as the current user's folder. This update also moves to a more robust system for copying data into the backup folder, with options set to minimize the amount of extra feedback shown in the console window while ensuring the best performance possible. Please note that user account backups can still take quite a bit of time.